# api/openapi.yaml
openapi: "3.1.0"
info:
  title: ISC Dashboard Backend API
  description: Backend API für das ISC Dashboard mit LDAP-Integration, Benutzer-, Widget- und News-Verwaltung
  version: "1.0.0"
servers:
  - url: http://127.0.0.1:8000
    description: Lokaler Development Server
paths:
...
  /api/v1/widgets/:
    get:
      tags:
        - widgets
      summary: Liste aller Widgets
      description: Ruft eine Liste aller Widgets ab mit optionaler Filterung nach Default-Status und Paginierung
      operationId: list_widgets
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: default
          in: query
          required: false
          schema:
            type: boolean
            nullable: true
          description: Filtert Widgets nach Default-Status (true = nur Default-Widgets, false = nur Nicht-Default-Widgets)
      responses:
        "200":
          description: Liste der Widgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WidgetRead"
    post:
      tags:
        - widgets
      summary: Neues Widget erstellen
      operationId: create_widget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WidgetCreate"
      responses:
        "201":
          description: Widget erstellt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WidgetRead"
    put:
      tags:
        - widgets
      summary: Bulk-Update von Widgets
      operationId: bulk_update_widgets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/WidgetUpdate"
      responses:
        "200":
          description: Widgets aktualisiert
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WidgetRead"
        "400":
          description: Fehler beim Update
  /api/v1/widgets/user-widgets/bulk:
    put:
      tags:
        - widgets
      summary: Bulk-Update von User-Widget-Konfigurationen
      description: Aktualisiert mehrere User-Widget-Einstellungen (Positionen, Sichtbarkeit) auf einmal. Wird verwendet, um Widget-Layout-Änderungen zu speichern. Erstellt neue Einträge, falls sie nicht existieren.
      operationId: bulk_update_user_widgets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/UserWidgetUpdate"
            example:
              - userId: "550e8400-e29b-41d4-a716-446655440000"
                widgetId: "660e8400-e29b-41d4-a716-446655440001"
                visible: true
                config:
                  x: 0
                  y: 1
              - userId: "550e8400-e29b-41d4-a716-446655440000"
                widgetId: "660e8400-e29b-41d4-a716-446655440002"
                visible: false
                config:
                  x: 1
                  y: 1
      responses:
        "200":
          description: User-Widgets aktualisiert
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserWidgetRead"
        "400":
          description: Fehler beim Update
  /api/v1/widgets/{user_id}:
    get:
      tags:
        - widgets
      summary: Widgets eines Users
      description: Ruft alle Widgets eines bestimmten Benutzers ab. Bei Erstanmeldung (keine User-Widgets vorhanden) werden automatisch alle Default-Widgets initialisiert.
      operationId: get_user_widgets
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: Liste der User-Widgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserWidgetRead"
  /api/v1/widget/{widget_id}:
    put:
      tags:
        - widgets
      summary: Einzelnes Widget aktualisieren
      operationId: update_widget
      parameters:
        - name: widget_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WidgetUpdate"
      responses:
        "200":
          description: Widget aktualisiert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WidgetRead"
        "404":
          description: Widget nicht gefunden
    delete:
      tags:
        - widgets
      summary: Widget löschen
      operationId: delete_widget
      parameters:
        - name: widget_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Widget gelöscht
        "404":
          description: Widget nicht gefunden
...
components:
  schemas:
...
    WidgetRead:
      type: object
      properties:
        widgetId:
          type: string
          format: uuid
        target:
          type: string
        icon:
          type: string
        title:
          type: string
        color:
          type: string
        default:
          type: boolean
          default: false
      required:
        - widgetId
        - target
        - icon
        - title
        - color
        - default
    WidgetCreate:
      type: object
      properties:
        target:
          type: string
        icon:
          type: string
        title:
          type: string
        color:
          type: string
        default:
          type: boolean
          default: false
      required:
        - target
        - icon
        - title
        - color
    WidgetUpdate:
      type: object
      properties:
        widgetId:
          type: string
          format: uuid
        target:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        default:
          type: boolean
          nullable: true
    UserWidgetRead:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        widgetId:
          type: string
          format: uuid
        target:
          type: string
        icon:
          type: string
        title:
          type: string
        color:
          type: string
        visible:
          type: boolean
        config:
          type: object
          description: Widget-Konfiguration (z.B. Position mit x und y Koordinaten)
          example:
            x: 0
            y: 1
      required:
        - userId
        - widgetId
        - target
        - icon
        - title
        - color
        - visible
        - config
    UserWidgetUpdate:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        widgetId:
          type: string
          format: uuid
        visible:
          type: boolean
          nullable: true
          description: Sichtbarkeit des Widgets (false = versteckt, nicht gelöscht)
        config:
          type: object
          nullable: true
          description: Widget-Konfiguration (Position, etc.)
          example:
            x: 0
            y: 1
      required:
        - userId
        - widgetId
...